+++ Pipfile
@@ -7,6 +7,7 @@ name = "pypi"
 flask = "*"
 mysql-connector-python = "*"
 flask-sqlalchemy = "*"
+flask-login = "*"
 
 [dev-packages]
