+++ app/sksk_ko/templates/auth/login.html
@@ -1,0 +1,19 @@
+{%extends "base.html" %}
+{% block body %}
+
+
+<div class="login-page">
+
+    <h2 class="login-header">ログイン</h2>
+
+    <div class="form">
+        <form action="{{url_for('auth.login')}}" method="post" class="login-form">
+            <input type="text" name="email" placeholder="メールアドレス" required class="form-control">
+            <input type="password" name="password" placeholder="パスワード" required class="form-control">
+            <button type="submit" class="btn btn-primary">ログイン</button>
+        </form>
+    </div>
+
+</div>
+
+{%endblock%}
