+++ app/sksk_ko/templates/base.html
@@ -30,7 +30,7 @@
                   <a class="nav-link" href="{{url_for('auth.signup')}}">ユーザー登録</a>
                 </li>
                 <li class="nav-item">
-                  <a class="nav-link" href="{{url_for('auth.signup')}}">ログイン</a>
+                  <a class="nav-link" href="{{url_for('auth.login')}}">ログイン</a>
                 </li>
             </ul>
           </div>
