+++ app/sksk_ko/templates/index.html
@@ -7,8 +7,4 @@
 
 <p>スコア</p>
 
-{% for grade in grades %}
-<p>{{grade.grade}}</p>
-{% endfor %}
-
 {%endblock%}
